Split VideoEmbed.jsx into YouTubeEmbed.jsx and InstagramEmbed.jsx
