+++ src/index.css
@@ -183,33 +183,6 @@
   border-radius: .25rem;
 }
 
-.video-callout {
-  position: relative;
-  width: 100%;
-  height: 0;
-  padding-bottom: 56.25%;
-}
-
-.insta-callout {
-  position: relative;
-  max-width: 100%;
-  height: calc(100vh - 1rem);
-}
-
-.video {
-  position: absolute;
-  top: 0;
-  left: 0;
-  width: 100%;
-  height: 100%;
-}
-
-.insta-video {
-  max-width: 100%;
-  padding-bottom: 1.25rem;
-  height: calc(100vh - 2rem);
-}
-
 .form-row div {
   margin-top: 1em;
 }
